+++ docs/changes.rst
@@ -5,4 +5,4 @@ Changes
 
 This page will show when and which updates have been applied to the API, like when new dataset has been added or if some of the API's functionality has changed. Big changes will also be annouced on Twitter at `@OpenRaad <https://twitter.com/OpenRaad>`__. So check back every now and then!
 
-* 2015-10-27: Launch of the Open Raadsinformatie API.
+* 2015-10-27: Launch of the Openwob API.